fix(audio): compile seamless track switch

+++ Sources/AetherEngine/AetherEngine.swift
@@ -3785,7 +3785,7 @@ public final class AetherEngine: ObservableObject {
                     self.activeAudioTrackIndex = index
                     self.activeAudioDecoder = Self.softwareAudioDecoderLabel(
                         audioTracks: self.audioTracks,
-                        activeIndex: index
+                        activeIndex: Int32(index)
                     )
                     self.activateRendererAudioSession(audioSourceStreamIndex: Int32(index))
                     EngineLog.emit(
